fix: lint errors (#586)
~~**Do not merge**~~
Issue #, if available:

*Description of changes:*
Fix all lint errors in windows. 
*Testing done:*
Yes. Unit test and golangci-lint.


- [X] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------
