+++ cmd/finch/lima_args_windows.go
@@ -9,12 +9,12 @@ func (nc *nerdctlCommand) GetLimaArgs() []string {
 	wd, err := nc.systemDeps.GetWd()
 	if err != nil {
 		nc.logger.Warnln("failed to get working directory, will default to user home with error %s", err.Error())
-		return append([]string{"shell", limaInstanceName, "sudo", "-E"})
+		return []string{"shell", limaInstanceName, "sudo", "-E"}
 	}
 	wslPath, err := convertToWSLPath(nc.systemDeps, wd)
 	if err != nil {
 		nc.logger.Warnln("failed to convert to WSL path, will default to user home with error %s", err.Error())
-		return append([]string{"shell", limaInstanceName, "sudo", "-E"})
+		return []string{"shell", limaInstanceName, "sudo", "-E"}
 	}
-	return append([]string{"shell", "--workdir", wslPath, limaInstanceName, "sudo", "-E"})
+	return []string{"shell", "--workdir", wslPath, limaInstanceName, "sudo", "-E"}
 }
