+++ e2e/vm/virtualization_framework_rosetta_darwin_test.go
@@ -1,5 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
+
 //go:build darwin
 
 package vm
