+++ pkg/config/validate_windows.go
